for now, specifing read-only on query suffices
