+++ src/Sushi.MicroORM/DataQuery.cs
@@ -58,7 +58,7 @@ namespace Sushi.MicroORM
         /// <summary>
         /// Gets or sets the connection string to use when executing this query. If left empty, the default connectionstring for <typeparamref name="T"/> will be used.
         /// </summary>
-        public string? ConnectionString { get; set; }
+        public ConnectionString? ConnectionString { get; set; }
 
         /// <summary>
         /// Gets or sets a value indicating if this query is a read-only statements. If true, execution can be optimized by using a read-only connection if available on the connection string.
