+++ tests/Sushi.MicroORM.UnitTests/ServiceCollectionExtensionsTest.cs
@@ -27,21 +27,6 @@ namespace Sushi.MicroORM.UnitTests
         }
 
         [Fact]
-        public void RegisterReadOnlyConnectorDependenciesTest()
-        {
-            string connectionString = "Server=.;Initial Catalog=db;User ID=user;Password=pass;";
-
-            IServiceCollection serviceCollection = new ServiceCollection();
-            serviceCollection.AddMicroORM(connectionString);
-            var serviceProvider = serviceCollection.BuildServiceProvider();
-
-            // create default connector
-            var connector = serviceProvider.GetService<IReadOnlyConnector<TestClass>>();
-
-            Assert.NotNull(connector);
-        }
-
-        [Fact]
         public void RegisterConnectorDependenciesTest_WithPrivateConstructor()
         {
             string connectionString = "Server=.;Initial Catalog=db;User ID=user;Password=pass;";
